+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+## Improved
+* game overview: improve how editions and purchases are displayed
+
 ## 1.5.1 / 2023-11-14 21:10+01:00
 
 ## Improved